refactor: NetworkManager 역할 분리
