+++ OpenMarket/OpenMarket/MainViewController.swift
@@ -65,15 +65,15 @@ final class MainViewController: UIViewController {
         loadProductListToCollectionView()
         print(collectionView.visibleCells)
         
-//        let product = Product(id: nil, vendorId: nil, vendorName: nil, name: "스톤", description: "스톤스톤", thumbnail: nil, currency: .KRW, price: 12345, bargainPrice: 1, discountedPrice: 1, stock: 123, createdAt: nil, issuedAt: nil, images: nil, vendor: nil, secret: "966j8xcwknjhh7wj")
-//
-//        let image = UIImage(systemName: "a")!
-//
-//        DispatchQueue.global().sync {
-//            manager.postProductLists(params: product, images: [image]) {
-//                print("?")
-//            }
-//        }
+        let product = Product(id: nil, vendorId: nil, vendorName: nil, name: "스톤", description: "스톤스톤", thumbnail: nil, currency: .KRW, price: 12345, bargainPrice: 1, discountedPrice: 1, stock: 123, createdAt: nil, issuedAt: nil, images: nil, vendor: nil, secret: "966j8xcwknjhh7wj")
+
+        let image = UIImage(systemName: "a")!
+
+        DispatchQueue.global().sync {
+            manager.postProductLists(params: product, images: [image]) {
+                print("?")
+            }
+        }
     }
     
     func setupNavBar() {
